+++ client/src/app/class/_services/dashboard.service.ts
@@ -738,32 +738,32 @@ export class DashboardService {
    * Queries allDocs to get docs for either attendance or score. Performs a searchDateRange search (the past month) by default.
    * @param type
    * @param currentClass
-   * @param reportDate. If null, calculates today's date and calculates a month previous for searchEndKey. If reportDate === '*',  does not add reportDate to the query and returns all docs for the type, currentClass, and curriculumLabel.
+   * @param startDate - If reportDate === '*',  does not add reportDate to the query and returns all docs for the type, currentClass, and curriculumLabel.
+   * @param endDate - If null, either calculates today's date and calculates a month previous for searchEndKey - or - uses startDate if searchDateRange is true.
    * @param curriculumLabel - may be empty if ignoreCurriculumsForTracking is set in the class-registration form (currentClass).
-   * @param randomId
-   * @param searchDateRange
+   * @param randomId - the auto-generated id for the class - currentClass.metadata?.randomId.
+   * @param searchDateRange - if true, endDate is set to one month before startDate.
    */
-  async searchDocs(type: string, currentClass, reportDate: string, curriculumLabel: string, randomId: string, searchDateRange: boolean = true) {
-    let endDate;
+  async searchDocs(type: string, currentClass, startDate: string, endDate: string, curriculumLabel: string, randomId: string, searchDateRange: boolean = true) {
     let wildcardSearchString = '\uffff'
     this.db = await this.getUserDB();
-    if (!reportDate) {
-      reportDate = DateTime.local().toISODate()
+    if (!startDate) {
+      startDate = DateTime.local().toISODate()
     }
-    if (reportDate !== '*') {
-      const lastMonth = DateTime.fromISO(reportDate).minus({ months: 1 }).toISODate()
-      endDate = searchDateRange ? lastMonth : reportDate
+    if (startDate !== '*' && !endDate) {
+      const lastMonth = DateTime.fromISO(startDate).minus({ months: 1 }).toISODate()
+      endDate = searchDateRange ? lastMonth : startDate
     }
     
     const grade = this.getValue('grade', currentClass)
     let searchStartKey: string, searchEndKey: string
     if (curriculumLabel) {
-      if (reportDate) {
-        if (reportDate === '*') {
+      if (startDate) {
+        if (startDate === '*') {
           searchStartKey = type + '-' + sanitize(grade.replace(/\s+/g, '')) + '-' + randomId + '-' + sanitize(curriculumLabel.replace(/\s+/g, ''))
           searchEndKey = type + '-' + sanitize(grade.replace(/\s+/g, '')) + '-' + randomId + '-' + sanitize(curriculumLabel.replace(/\s+/g, ''))
         } else {
-          searchStartKey = type + '-' + sanitize(grade.replace(/\s+/g, '')) + '-' + randomId + '-' + sanitize(curriculumLabel.replace(/\s+/g, '')) + '-' + reportDate
+          searchStartKey = type + '-' + sanitize(grade.replace(/\s+/g, '')) + '-' + randomId + '-' + sanitize(curriculumLabel.replace(/\s+/g, '')) + '-' + startDate
           searchEndKey = type + '-' + sanitize(grade.replace(/\s+/g, '')) + '-' + randomId + '-' + sanitize(curriculumLabel.replace(/\s+/g, '')) + '-' + endDate
         }
         
@@ -772,12 +772,12 @@ export class DashboardService {
         searchEndKey = type + '-' + sanitize(grade.replace(/\s+/g, '')) + '-' + randomId + '-' + sanitize(curriculumLabel.replace(/\s+/g, ''))
       }
     } else {
-      if (reportDate) {
-        if (reportDate === '*') {
+      if (startDate) {
+        if (startDate === '*') {
           searchStartKey = type + '-' + sanitize(grade.replace(/\s+/g, '')) + '-' + randomId
           searchEndKey = type + '-' + sanitize(grade.replace(/\s+/g, '')) + '-' + randomId
         } else {
-          searchStartKey = type + '-' + sanitize(grade.replace(/\s+/g, '')) + '-' + randomId + '-' + reportDate
+          searchStartKey = type + '-' + sanitize(grade.replace(/\s+/g, '')) + '-' + randomId + '-' + startDate
           searchEndKey = type + '-' + sanitize(grade.replace(/\s+/g, '')) + '-' + randomId + '-' + endDate
         }
       } else {
@@ -790,7 +790,7 @@ export class DashboardService {
       endkey: searchEndKey + wildcardSearchString,
       include_docs: true
     }
-    if (reportDate) {
+    if (startDate) {
       options['startkey'] = searchStartKey +  wildcardSearchString
       options['endkey'] = searchEndKey
       options['descending'] = true