+++ client/src/app/class/attendance/behavior-check/behavior-check.component.ts
@@ -103,7 +103,7 @@ export class BehaviorCheckComponent implements OnInit {
 
     let currentBehaviorReport, savedBehaviorList = null;
     try {
-      const docArray = await this.dashboardService.searchDocs(type, currentClass, reportDate, curriculumLabel, randomId, false)
+      const docArray = await this.dashboardService.searchDocs(type, currentClass, reportDate, null, curriculumLabel, randomId, false)
       currentBehaviorReport = docArray? docArray[0]?.doc : null
       savedBehaviorList = currentBehaviorReport?.studentBehaviorList
     } catch (e) {